Simplify modal eval to a single file
