--- swebench/harness/modal_eval/__init__.py
@@ -1,8 +1,0 @@
-from swebench.harness.modal_eval.run_evaluation_modal import run_instances_modal
-from swebench.harness.modal_eval.utils import validate_modal_credentials
-
-
-__all__ = [
-    "run_instances_modal",
-    "validate_modal_credentials",
-]
